Fix structural guardrail failure by splitting list operation modules
